Retain attachment dir indexes
When running the repair_settings with attachments directories
with non continous indexes, the indexes was reset to a continous
index. If one of the reindexed directories was selected as current
attachment directory, it would point to the wrong, or non existing
attachment directory index. Fix this by passing and using the current
indexes.

Non coninous indexes can happen if an attachment directory is removed
that is not the last.

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>